Updated Mood Bot
Made changes as discussed, including:
- using doc strings for comments
- lines 58-71: use a dictionary for checking the day, mapping day of the week to an integer.
- line 36: make day regex global
- combine for loops in add/del functions
- use longer variable names consistently
 --> switch d to diary
 --> switch t to time_input and time_inputs
- used fixed strings or functions for repeated messages
- defined custom exceptions (will add in another PR).
